I tried matching the type of the float_input.go file, but that didn't work... oh well

+++ section2/float_output.go
@@ -9,10 +9,10 @@ func init() {
 }
 
 //export float_output
-func float_output() float32 {
+func float_output() float64 {
 	// Look into seeding a random generator
 	// figure out how to get random number.
-	return (rand.Float32() + float32(1)) * float32(rand.Intn(1000))
+	return (rand.Float64() + float64(1)) * float64(rand.Intn(1000))
 }
 
 func main() {}
